feat : receipt 조회 시 id 반환하지 않고 Todo 정보 반환

+++ src/main/java/com/sprint/dailyreceipt/domain/receipt/api/model/ReceiptRegisterRequest.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 public class ReceiptRegisterRequest {
 
-    private List<Integer> todoIds;
+    private List<Long> todoIds;
 
     private boolean pinned;
 
@@ -21,7 +21,7 @@ public class ReceiptRegisterRequest {
     private String name;
 
     @Builder
-    public ReceiptRegisterRequest(List<Integer> todoIds, boolean pinned, String famousSaying, String name) {
+    public ReceiptRegisterRequest(List<Long> todoIds, boolean pinned, String famousSaying, String name) {
         this.todoIds = todoIds;
         this.pinned = pinned;
         this.famousSaying = famousSaying;
